Improve .vimrc
I don't miss whatever plugins I was using with pathogen so that line is
commented out to avoid startup noise.

The statusline has the buffer number and modified status added. Not sure
how useful the buffer number is but will keep it for now.

The diffopt is useful for the :diffs command, which is split vertically.
The other two options are defaults which are kept without examination.

We eliminate windowsfont as we thankfully don't need to endure that any
longer.

Finally, typescript files are set to javascriptreact instead, as that
builtin filetype has proper syntax highlighting etc. The indentation and
such matches work repo style.