Update diffopt, set cursorline, save search trick
diffopt=internal causes errors on the macOS builtin vim, despite
featuring in the documentation, but since this is a default option we
can remove it for now and see if it causes trouble on other OSes.

Weirdly, I tried `set diffopt+=vertical` but it didn't work, even though
`:set diffopt+=vertical` works interactively. Never mind.

The cursorline feature is great, thanks reddit. Definitely a keeper.

The final tip is saved in case the reddit post disappears and for ease
of use - I'm not 100% sure if I want to keep it or not so it stays
commented for now.