Change clipboard paste for Ubuntu 21.10
For some reason the version of vim that comes with Ubuntu doesn't have
the xterm_clipboard option compiled in, so as per the documentation
(nice) these bindings just work with the unnamed register.

Rather than messing around compiling or looking for a version which does
have this, it's easier to just tweak the mappings and move on. Yanking
to clipboard doesn't seem to come up often and the mapping feels
complicated so I'll do without for now and see if it becomes an issue.